sys/dev/disk/dm: Remove unnecessary cast
that was added in 66fee7c.

+++ sys/dev/disk/dm/targets/delay/dm_target_delay.c
@@ -377,7 +377,7 @@ _makeudev(struct dm_delay_info *di)
 	KKASSERT(error == 0);
 
 	ret = makeudev(va.va_rmajor, va.va_rminor);
-	return (uint64_t)ret;
+	return ret;
 }
 
 static int
